custom pages added for check list and added new custom label for tablet ui fix
