+++ apps/health_campaign_field_worker_app/lib/pages/beneficiary/custom_deliver_intervention.dart
@@ -705,8 +705,9 @@ class CustomDeliverInterventionPageState
                                                             icon: Icons.qr_code,
                                                             label: localizations
                                                                 .translate(
-                                                              i18.individualDetails
-                                                                  .linkVoucherToIndividual,
+                                                              i18Local
+                                                                  .deliverIntervention
+                                                                  .scanningITNs,
                                                             ),
                                                           ),
                                                   ),
