+++ apps/health_campaign_field_worker_app/pubspec.yaml
@@ -1,7 +1,7 @@
 name: health_campaign_field_worker_app
 description: A new Flutter project.
 publish_to: "none"
-version: 1.6.13+9
+version: 1.6.14+9
 
 environment:
   sdk: ">=3.0.0 <=4.0.0"
